2023. 5. 17. (WED)

+++ src/main/kotlin/com/sheenhm/cabbagemarket/repository/UsedItemRepository.kt
@@ -3,13 +3,14 @@ package com.sheenhm.cabbagemarket.repository
 import jakarta.persistence.*
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
+import java.util.*
 
 @Entity
 @Table(name = "used_items")
 data class UsedItem(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long,
+    val id: Int,
 
     @Column(nullable = false)
     val title: String,
@@ -18,15 +19,15 @@ data class UsedItem(
     val imageUrl: String? = null,
 
     @Column(nullable = false)
-    val price: Long,
+    val price: Int,
 
     @Column(name = "seller_id")
-    val sellerId: Long
+    val sellerId: Int
 ) {
     constructor() : this(0, "", null, 0, 0)
 }
 
 @Repository
-interface UsedItemRepository: JpaRepository<UsedItem, Long> {
+interface UsedItemRepository: JpaRepository<UsedItem, Int> {
     fun findBySellerId(sellerId: Long): List<UsedItem>
 }
